Module mango_v4::accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is a Pubkey. This is useful for specifying accounts for a client.

Structs

AccountBuybackFeesWithMngo

Generated client accounts for AccountBuybackFeesWithMngo.

AccountClose

Generated client accounts for AccountClose.

AccountCreate

Generated client accounts for AccountCreate.

AccountEdit

Generated client accounts for AccountEdit.

AccountExpand

Generated client accounts for AccountExpand.

AccountToggleFreeze

Generated client accounts for AccountToggleFreeze.

AltExtend

Generated client accounts for AltExtend.

AltSet

Generated client accounts for AltSet.

Benchmark

Generated client accounts for Benchmark.

ComputeAccountData

Generated client accounts for ComputeAccountData.

FlashLoanBegin

Generated client accounts for FlashLoanBegin.

FlashLoanEnd

Generated client accounts for FlashLoanEnd.

GroupClose

Generated client accounts for GroupClose.

GroupCreate

Generated client accounts for GroupCreate.

GroupEdit

Generated client accounts for GroupEdit.

HealthRegionBegin

Generated client accounts for HealthRegionBegin.

HealthRegionEnd

Generated client accounts for HealthRegionEnd.

IxGateSet

Generated client accounts for IxGateSet.

PerpCancelAllOrders

Generated client accounts for PerpCancelAllOrders.

PerpCancelAllOrdersBySide

Generated client accounts for PerpCancelAllOrdersBySide.

PerpCancelOrder

Generated client accounts for PerpCancelOrder.

PerpCancelOrderByClientOrderId

Generated client accounts for PerpCancelOrderByClientOrderId.

PerpCloseMarket

Generated client accounts for PerpCloseMarket.

PerpConsumeEvents

Generated client accounts for PerpConsumeEvents.

PerpCreateMarket

Generated client accounts for PerpCreateMarket.

PerpDeactivatePosition

Generated client accounts for PerpDeactivatePosition.

PerpEditMarket

Generated client accounts for PerpEditMarket.

PerpLiqBaseOrPositivePnl

Generated client accounts for PerpLiqBaseOrPositivePnl.

PerpLiqForceCancelOrders

Generated client accounts for PerpLiqForceCancelOrders.

PerpLiqNegativePnlOrBankruptcy

Generated client accounts for PerpLiqNegativePnlOrBankruptcy.

PerpPlaceOrder

Generated client accounts for PerpPlaceOrder.

PerpSettleFees

Generated client accounts for PerpSettleFees.

PerpSettlePnl

Generated client accounts for PerpSettlePnl.

PerpUpdateFunding

Generated client accounts for PerpUpdateFunding.

Serum3CancelAllOrders

Generated client accounts for Serum3CancelAllOrders.

Serum3CancelOrder

Generated client accounts for Serum3CancelOrder.

Serum3CloseOpenOrders

Generated client accounts for Serum3CloseOpenOrders.

Serum3CreateOpenOrders

Generated client accounts for Serum3CreateOpenOrders.

Serum3DeregisterMarket

Generated client accounts for Serum3DeregisterMarket.

Serum3EditMarket

Generated client accounts for Serum3EditMarket.

Serum3LiqForceCancelOrders

Generated client accounts for Serum3LiqForceCancelOrders.

Serum3PlaceOrder

Generated client accounts for Serum3PlaceOrder.

Serum3RegisterMarket

Generated client accounts for Serum3RegisterMarket.

Serum3SettleFunds

Generated client accounts for Serum3SettleFunds.

Serum3SettleFundsV2

Generated client accounts for Serum3SettleFundsV2.

Serum3SettleFundsV2Extra

Generated client accounts for Serum3SettleFundsV2Extra.

StubOracleClose

Generated client accounts for StubOracleClose.

StubOracleCreate

Generated client accounts for StubOracleCreate.

StubOracleSet

Generated client accounts for StubOracleSet.

TokenAddBank

Generated client accounts for TokenAddBank.

TokenDeposit

Generated client accounts for TokenDeposit.

TokenDepositIntoExisting

Generated client accounts for TokenDepositIntoExisting.

TokenDeregister

Generated client accounts for TokenDeregister.

TokenEdit

Generated client accounts for TokenEdit.

TokenForceCloseBorrowsWithToken

Generated client accounts for TokenForceCloseBorrowsWithToken.

TokenLiqBankruptcy

Generated client accounts for TokenLiqBankruptcy.

TokenLiqWithToken

Generated client accounts for TokenLiqWithToken.

TokenRegister

Generated client accounts for TokenRegister.

TokenRegisterTrustless

Generated client accounts for TokenRegisterTrustless.

TokenUpdateIndexAndRate

Generated client accounts for TokenUpdateIndexAndRate.

TokenWithdraw

Generated client accounts for TokenWithdraw.