Module mango_v4::instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

Modules

state

Instruction struct definitions for #[state] methods.

Structs

AccountBuybackFeesWithMngo

Instruction.

AccountClose

Instruction.

AccountCreate

Instruction.

AccountEdit

Instruction.

AccountExpand

Instruction.

AccountToggleFreeze

Instruction.

AltExtend

Instruction.

AltSet

Instruction.

Benchmark

Instruction.

ComputeAccountData

Instruction.

FlashLoanBegin

Instruction.

FlashLoanEnd

Instruction.

FlashLoanEndV2

Instruction.

GroupClose

Instruction.

GroupCreate

Instruction.

GroupEdit

Instruction.

HealthRegionBegin

Instruction.

HealthRegionEnd

Instruction.

IxGateSet

Instruction.

LiqTokenBankruptcy

Instruction.

LiqTokenWithToken

Instruction.

PerpCancelAllOrders

Instruction.

PerpCancelAllOrdersBySide

Instruction.

PerpCancelOrder

Instruction.

PerpCancelOrderByClientOrderId

Instruction.

PerpCloseMarket

Instruction.

PerpConsumeEvents

Instruction.

PerpCreateMarket

Instruction.

PerpDeactivatePosition

Instruction.

PerpEditMarket

Instruction.

PerpLiqBaseOrPositivePnl

Instruction.

PerpLiqForceCancelOrders

Instruction.

PerpLiqNegativePnlOrBankruptcy

Instruction.

PerpPlaceOrder

Instruction.

PerpPlaceOrderPegged

Instruction.

PerpSettleFees

Instruction.

PerpSettlePnl

Instruction.

PerpUpdateFunding

Instruction.

Serum3CancelAllOrders

Instruction.

Serum3CancelOrder

Instruction.

Serum3CloseOpenOrders

Instruction.

Serum3CreateOpenOrders

Instruction.

Serum3DeregisterMarket

Instruction.

Serum3EditMarket

Instruction.

Serum3LiqForceCancelOrders

Instruction.

Serum3PlaceOrder

Instruction.

Serum3RegisterMarket

Instruction.

Serum3SettleFunds

Instruction.

Serum3SettleFundsV2

Instruction.

StubOracleClose

Instruction.

StubOracleCreate

Instruction.

StubOracleSet

Instruction.

TokenAddBank

Instruction.

TokenDeposit

Instruction.

TokenDepositIntoExisting

Instruction.

TokenDeregister

Instruction.

TokenEdit

Instruction.

TokenForceCloseBorrowsWithToken

Instruction.

TokenLiqBankruptcy

Instruction.

TokenLiqWithToken

Instruction.

TokenRegister

Instruction.

TokenRegisterTrustless

Instruction.

TokenUpdateIndexAndRate

Instruction.

TokenWithdraw

Instruction.