Module mango_v4::instructions

Modules

compute_budget

jupiter_mainnet_3

jupiter_mainnet_4

Structs

AccountClose

AccountCreate

AccountEdit

AccountExpand

AccountToggleFreeze

AltExtend

AltSet

Benchmark

ComputeAccountData

FlashLoanBegin

Sets up mango vaults for flash loan

FlashLoanEnd

Finalizes a flash loan

GroupClose

GroupCreate

GroupEdit

HealthRegionBegin

Sets up for a health region

HealthRegionEnd

Ends a health region.

InterestRateParams

IxGateSet

OODifference

OpenOrdersSlim

For loan origination fees bookkeeping purposes

PerpCancelAllOrders

PerpCancelAllOrdersBySide

PerpCancelOrder

PerpCancelOrderByClientOrderId

PerpCloseMarket

PerpConsumeEvents

PerpCreateMarket

PerpDeactivatePosition

PerpEditMarket

PerpLiqBaseOrPositivePnl

PerpLiqForceCancelOrders

PerpLiqNegativePnlOrBankruptcy

PerpPlaceOrder

PerpSettleFees

PerpSettlePnl

PerpUpdateFunding

Serum3CancelAllOrders

Serum3CancelOrder

Serum3CloseOpenOrders

Serum3CreateOpenOrders

Serum3DeregisterMarket

Serum3EditMarket

Serum3LiqForceCancelOrders

Serum3PlaceOrder

Serum3RegisterMarket

Serum3SettleFunds

StubOracleClose

StubOracleCreate

StubOracleSet

TokenAddBank

TokenDeposit

TokenDepositIntoExisting

TokenDeregister

In addition to these accounts, there must be remaining_accounts: all n pairs of bank and its corresponding vault account for a token

TokenEdit

Changes a token’s parameters.

TokenLiqBankruptcy

TokenLiqWithToken

TokenRegister

TokenRegisterTrustless

TokenUpdateIndexAndRate

Updates token interest and interest rates.

TokenWithdraw

VaultDifference

Enums

FlashLoanType

Serum3OrderType

Serum3SelfTradeBehavior

Copy paste a bunch of enums so that we could AnchorSerialize & AnchorDeserialize them

Serum3Side

Constants

INDEX_START

Traits

OpenOrdersAmounts

Functions

account_close

account_create

account_edit

account_expand

account_toggle_freeze

alt_extend

Add addresses to a registered lookup table

alt_set

Register or replace an address lookup table

apply_vault_difference

Called in settle_funds, place_order, liq_force_cancel to adjust token positions after changing the vault balances Also logs changes to token balances

benchmark

charge_loan_origination_fees

compute_account_data

compute_settle_fee

division_i32

division_i64

division_i80f48

division_i80f48_30bit

division_i80f48_f64

division_i128

division_u32

division_u80f48

flash_loan_begin

The loan_amounts argument lists the amount to be loaned from each bank/vault and the order matches the order of bank accounts.

flash_loan_end

group_close

group_create

group_edit

health_region_begin

health_region_end

i80f48_to_f64

ix_gate_set

mul_f64

mul_i80f48

perp_cancel_all_orders

perp_cancel_all_orders_by_side

perp_cancel_order

perp_cancel_order_by_client_order_id

perp_close_market

perp_consume_events

perp_create_market

perp_deactivate_position

perp_edit_market

perp_liq_base_or_positive_pnl

This instruction deals with increasing health by:

perp_liq_force_cancel_orders

perp_liq_negative_pnl_or_bankruptcy

perp_place_order

perp_settle_fees

perp_settle_pnl

perp_update_funding

serum3_cancel_all_orders

serum3_cancel_order

serum3_close_open_orders

serum3_create_open_orders

serum3_deregister_market

serum3_edit_market

serum3_liq_force_cancel_orders

serum3_place_order

serum3_register_market

serum3_settle_funds

Settling means moving free funds from the serum3 open orders account back into the mango account wallet.

stub_oracle_close

stub_oracle_create

stub_oracle_set

token_add_bank

token_deposit

token_deposit_into_existing

token_deregister

token_edit

token_liq_bankruptcy

token_liq_with_token

token_register

token_register_trustless

token_update_index_and_rate

token_withdraw