Struct mango_v4::state::BookSide

#[repr(C)]
pub struct BookSide {
    pub roots: [OrderTreeRoot; 2],
    pub reserved_roots: [OrderTreeRoot; 4],
    pub reserved: [u8; 256],
    pub nodes: OrderTreeNodes,
}

Fields
roots: [OrderTreeRoot; 2]reserved_roots: [OrderTreeRoot; 4]reserved: [u8; 256]nodes: OrderTreeNodes
Implementations

impl BookSide

pub fn iter_valid(
&self,
now_ts: u64,
oracle_price_lots: i64
) -> impl Iterator<Item = BookSideIterItem<'_>>

Iterate over all entries in the book filtering out invalid orders

smallest to highest for asks highest to smallest for bids

pub fn iter_all_including_invalid(
&self,
now_ts: u64,
oracle_price_lots: i64
) -> BookSideIter<'_>

Iterate over all entries, including invalid orders

pub fn node(&self, handle: NodeHandle) -> Option<&AnyNode>

pub fn node_mut(&mut self, handle: NodeHandle) -> Option<&mut AnyNode>

pub fn root(&self, component: BookSideOrderTree) -> &OrderTreeRoot

pub fn root_mut(&mut self, component: BookSideOrderTree) -> &mut OrderTreeRoot

pub fn is_full(&self) -> bool

pub fn insert_leaf(
&mut self,
component: BookSideOrderTree,
new_leaf: &LeafNode
) -> Result<(NodeHandle, Option<LeafNode>)>

pub fn remove_worst(
&mut self,
now_ts: u64,
oracle_price_lots: i64
) -> Option<(LeafNode, i64)>

Remove the overall worst-price order.

pub fn remove_one_expired(
&mut self,
component: BookSideOrderTree,
now_ts: u64
) -> Option<LeafNode>

Remove the order with the lowest expiry timestamp in the component, if that’s < now_ts. If there is none, try to remove the lowest expiry one from the other component.

pub fn remove_by_key(
&mut self,
component: BookSideOrderTree,
search_key: u128
) -> Option<LeafNode>

pub fn side(&self) -> Side

pub fn quantity_at_price(
&self,
limit_price_lots: i64,
now_ts: u64,
oracle_price_lots: i64
) -> i64

Return the quantity of orders that can be matched by an order at limit_price_lots

pub fn best_price(&self, now_ts: u64, oracle_price_lots: i64) -> Option<i64>

Return the price of the order closest to the spread

pub fn impact_price(
&self,
quantity: i64,
now_ts: u64,
oracle_price_lots: i64
) -> Option<i64>

Walk up the book quantity units and return the price at that level. If quantity units not on book, return None