Struct mango_v4::state::MangoAccount

pub struct MangoAccount {

    pub group: Pubkey,
    pub owner: Pubkey,
    pub name: [u8; 32],
    pub delegate: Pubkey,
    pub account_num: u32,
    pub being_liquidated: u8,
    pub in_health_region: u8,
    pub bump: u8,
    pub padding: [u8; 1],
    pub net_deposits: i64,
    pub perp_spot_transfers: i64,
    pub health_region_begin_init_health: i64,
    pub frozen_until: u64,
    pub buyback_fees_accrued_current: u64,
    pub buyback_fees_accrued_previous: u64,
    pub buyback_fees_expiry_timestamp: u64,
    pub next_token_conditional_swap_id: u64,
    pub reserved: [u8; 200],
    pub header_version: u8,
    pub padding3: [u8; 7],
    pub padding4: u32,
    pub tokens: Vec<TokenPosition>,
    pub padding5: u32,
    pub serum3: Vec<Serum3Orders>,
    pub padding6: u32,
    pub perps: Vec<PerpPosition>,
    pub padding7: u32,
    pub perp_open_orders: Vec<PerpOpenOrder>,

}

Fields
group: Pubkeyowner: Pubkeyname: [u8; 32]delegate: Pubkeyaccount_num: u32being_liquidated: u8

Tracks that this account should be liquidated until init_health >= 0.

Normally accounts can not be liquidated while maint_health >= 0. But when an account reaches maint_health < 0, liquidators will call a liquidation instruction and thereby set this flag. Now the account may be liquidated until init_health >= 0.

Many actions should be disabled while the account is being liquidated, even if its maint health has recovered to positive. Creating new open orders would, for example, confuse liquidators.

in_health_region: u8

The account is currently inside a health region marked by HealthRegionBegin…HealthRegionEnd.

Must never be set after a transaction ends.

bump: u8padding: [u8; 1]net_deposits: i64perp_spot_transfers: i64health_region_begin_init_health: i64

Init health as calculated during HealthReginBegin, rounded up.

frozen_until: u64buyback_fees_accrued_current: u64

Fees usable with the “fees buyback” feature. This tracks the ones that accrued in the current expiry interval.

buyback_fees_accrued_previous: u64

Fees buyback amount from the previous expiry interval.

buyback_fees_expiry_timestamp: u64

End timestamp of the current expiry interval of the buyback fees amount.

next_token_conditional_swap_id: u64

Next id to use when adding a token condition swap

reserved: [u8; 200]header_version: u8padding3: [u8; 7]padding4: u32tokens: Vec<TokenPosition>padding5: u32serum3: Vec<Serum3Orders>padding6: u32perps: Vec<PerpPosition>padding7: u32perp_open_orders: Vec<PerpOpenOrder>
Implementations

impl MangoAccount

pub fn default_for_tests() -> Self

pub fn space(
token_count: u8,
serum3_count: u8,
perp_count: u8,
perp_oo_count: u8,
token_conditional_swap_count: u8
) -> usize

Number of bytes needed for the MangoAccount, including the discriminator

pub fn dynamic_token_vec_offset() -> usize

pub fn dynamic_serum3_vec_offset(token_count: u8) -> usize

pub fn dynamic_perp_vec_offset(token_count: u8, serum3_count: u8) -> usize

pub fn dynamic_perp_oo_vec_offset(
token_count: u8,
serum3_count: u8,
perp_count: u8
) -> usize

pub fn dynamic_token_conditional_swap_vec_offset(
token_count: u8,
serum3_count: u8,
perp_count: u8,
perp_oo_count: u8
) -> usize

pub fn dynamic_reserved_bytes_offset(
token_count: u8,
serum3_count: u8,
perp_count: u8,
perp_oo_count: u8,
token_conditional_swap_count: u8
) -> usize

pub fn dynamic_size(
token_count: u8,
serum3_count: u8,
perp_count: u8,
perp_oo_count: u8,
token_conditional_swap_count: u8
) -> usize