Trait mango_v4::state::DynamicHeader

pub trait DynamicHeader: Sized {
    fn from_bytes(dynamic_data: &[u8]) -> Result<Self>;
    fn initialize(dynamic_data: &mut [u8]) -> Result<()>;
}

Header is created by scanning and parsing the dynamic portion of the account. This stores useful information e.g. offsets to easily seek into dynamic content.

Required Methods

fn from_bytes(dynamic_data: &[u8]) -> Result<Self>

Builds header by scanning and parsing the dynamic portion of the account.

fn initialize(dynamic_data: &mut [u8]) -> Result<()>

initializes a header on the dynamic portion of a new account