Module mango_v4::accounts_ix

Modules

compute_budget

jupiter_mainnet_3

jupiter_mainnet_4

Structs

AccountBuybackFeesWithMngo

AccountClose

AccountCreate

AccountCreateV2

AccountEdit

AccountExpand

AccountSizeMigration

AccountToggleFreeze

AdminPerpWithdrawFees

AdminTokenWithdrawFees

AltExtend

AltSet

Benchmark

ComputeAccountData

FlashLoanBegin

Sets up mango vaults for flash loan

FlashLoanEnd

Finalizes a flash loan

GroupClose

GroupCreate

GroupEdit

GroupWithdrawInsuranceFund

HealthRegionBegin

Sets up for a health region

HealthRegionEnd

Ends a health region.

InterestRateParams

IxGateSet

OpenbookV2CancelOrder

OpenbookV2CloseOpenOrders

OpenbookV2CreateOpenOrders

OpenbookV2DeregisterMarket

OpenbookV2EditMarket

OpenbookV2LiqForceCancelOrders

OpenbookV2PlaceOrder

OpenbookV2PlaceTakeOrder

OpenbookV2RegisterMarket

OpenbookV2SettleFunds

PerpCancelAllOrders

PerpCancelAllOrdersBySide

PerpCancelOrder

PerpCancelOrderByClientOrderId

PerpCloseMarket

PerpConsumeEvents

PerpCreateMarket

PerpDeactivatePosition

PerpEditMarket

PerpForceClosePosition

PerpLiqBaseOrPositivePnl

PerpLiqForceCancelOrders

PerpLiqNegativePnlOrBankruptcy

PerpLiqNegativePnlOrBankruptcyV2

PerpPlaceOrder

PerpSettleFees

PerpSettlePnl

PerpUpdateFunding

Serum3CancelAllOrders

Serum3CancelOrder

Serum3CloseOpenOrders

Serum3CreateOpenOrders

Serum3DeregisterMarket

Serum3EditMarket

Serum3LiqForceCancelOrders

Serum3PlaceOrder

Serum3RegisterMarket

Serum3SettleFunds

Serum3SettleFundsV2

Serum3SettleFundsV2Extra

StubOracleClose

StubOracleCreate

StubOracleSet

TokenAddBank

TokenConditionalSwapCancel

TokenConditionalSwapCreate

TokenConditionalSwapTrigger

TokenDeposit

TokenDepositIntoExisting

TokenDeregister

In addition to these accounts, there must be remaining_accounts: all n pairs of bank and its corresponding vault account for a token

TokenEdit

Changes a token’s parameters.

TokenForceCloseBorrowsWithToken

TokenLiqBankruptcy

TokenLiqWithToken

TokenRegister

TokenRegisterTrustless

TokenUpdateIndexAndRate

Updates token interest and interest rates.

TokenWithdraw

Enums

FlashLoanType

Serum3OrderType

Serum3SelfTradeBehavior

Copy paste a bunch of enums so that we could AnchorSerialize & AnchorDeserialize them

Serum3Side