Module mango_v4::instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

Structs

AccountBuybackFeesWithMngo

Instruction.

AccountClose

Instruction.

AccountCreate

Instruction.

AccountCreateV2

Instruction.

AccountEdit

Instruction.

AccountExpand

Instruction.

AccountExpandV2

Instruction.

AccountSizeMigration

Instruction.

AccountToggleFreeze

Instruction.

AdminPerpWithdrawFees

Instruction.

AdminTokenWithdrawFees

Instruction.

AltExtend

Instruction.

AltSet

Instruction.

Benchmark

Instruction.

ComputeAccountData

Instruction.

FlashLoanBegin

Instruction.

FlashLoanEnd

Instruction.

FlashLoanEndV2

Instruction.

GroupClose

Instruction.

GroupCreate

Instruction.

GroupEdit

Instruction.

GroupWithdrawInsuranceFund

Instruction.

HealthRegionBegin

Instruction.

HealthRegionEnd

Instruction.

IxGateSet

Instruction.

LiqTokenBankruptcy

Instruction.

LiqTokenWithToken

Instruction.

OpenbookV2CancelAllOrders

Instruction.

OpenbookV2CancelOrder

Instruction.

OpenbookV2CloseOpenOrders

Instruction.

OpenbookV2CreateOpenOrders

Instruction.

OpenbookV2DeregisterMarket

Instruction.

OpenbookV2EditMarket

Instruction.

OpenbookV2LiqForceCancelOrders

Instruction.

OpenbookV2PlaceOrder

Instruction.

OpenbookV2PlaceTakerOrder

Instruction.

OpenbookV2RegisterMarket

Instruction.

OpenbookV2SettleFunds

Instruction.

PerpCancelAllOrders

Instruction.

PerpCancelAllOrdersBySide

Instruction.

PerpCancelOrder

Instruction.

PerpCancelOrderByClientOrderId

Instruction.

PerpCloseMarket

Instruction.

PerpConsumeEvents

Instruction.

PerpCreateMarket

Instruction.

PerpDeactivatePosition

Instruction.

PerpEditMarket

Instruction.

PerpForceClosePosition

Instruction.

PerpLiqBaseOrPositivePnl

Instruction.

PerpLiqForceCancelOrders

Instruction.

PerpLiqNegativePnlOrBankruptcy

Instruction.

PerpLiqNegativePnlOrBankruptcyV2

Instruction.

PerpPlaceOrder

Instruction.

PerpPlaceOrderPegged

Instruction.

PerpPlaceOrderPeggedV2

Instruction.

PerpPlaceOrderV2

Instruction.

PerpSettleFees

Instruction.

PerpSettlePnl

Instruction.

PerpUpdateFunding

Instruction.

Serum3CancelAllOrders

Instruction.

Serum3CancelOrder

Instruction.

Serum3CloseOpenOrders

Instruction.

Serum3CreateOpenOrders

Instruction.

Serum3DeregisterMarket

Instruction.

Serum3EditMarket

Instruction.

Serum3LiqForceCancelOrders

Instruction.

Serum3PlaceOrder

Instruction.

Serum3RegisterMarket

Instruction.

Serum3SettleFunds

Instruction.

Serum3SettleFundsV2

Instruction.

StubOracleClose

Instruction.

StubOracleCreate

Instruction.

StubOracleSet

Instruction.

StubOracleSetTest

Instruction.

TokenAddBank

Instruction.

TokenConditionalSwapCancel

Instruction.

TokenConditionalSwapCreate

Instruction.

TokenConditionalSwapCreateV2

Instruction.

TokenConditionalSwapTrigger

Instruction.

TokenDeposit

Instruction.

TokenDepositIntoExisting

Instruction.

TokenDeregister

Instruction.

TokenEdit

Instruction.

TokenForceCloseBorrowsWithToken

Instruction.

TokenLiqBankruptcy

Instruction.

TokenLiqWithToken

Instruction.

TokenRegister

Instruction.

TokenRegisterTrustless

Instruction.

TokenUpdateIndexAndRate

Instruction.

TokenWithdraw

Instruction.