Module mango_v4::mango_v4

Functions

account_buyback_fees_with_mngo

account_close

account_create

account_create_v2

account_edit

account_expand

account_expand_v2

account_size_migration

account_toggle_freeze

admin_perp_withdraw_fees

admin_token_withdraw_fees

alt_extend

alt_set

benchmark

benchmark

compute_account_data

Warning, this instruction is for testing purposes only!

flash_loan_begin

flash_loan_end

flash_loan_end_v2

group_close

group_create

group_edit

group_withdraw_insurance_fund

health_region_begin

health_region_end

ix_gate_set

liq_token_bankruptcy

liq_token_with_token

openbook_v2_cancel_all_orders

openbook_v2_cancel_order

openbook_v2_close_open_orders

openbook_v2_create_open_orders

openbook_v2_deregister_market

openbook_v2_edit_market

openbook_v2_liq_force_cancel_orders

openbook_v2_place_order

openbook_v2_place_taker_order

openbook_v2_register_market

OpenbookV2

openbook_v2_settle_funds

perp_cancel_all_orders

perp_cancel_all_orders_by_side

perp_cancel_order

perp_cancel_order_by_client_order_id

perp_close_market

perp_consume_events

perp_create_market

Perps

perp_deactivate_position

perp_edit_market

perp_force_close_position

perp_liq_base_or_positive_pnl

perp_liq_force_cancel_orders

perp_liq_negative_pnl_or_bankruptcy

perp_liq_negative_pnl_or_bankruptcy_v2

perp_place_order

perp_place_order_pegged

perp_place_order_pegged_v2

perp_place_order_v2

perp_settle_fees

perp_settle_pnl

perp_update_funding

serum3_cancel_all_orders

serum3_cancel_order

serum3_close_open_orders

serum3_create_open_orders

serum3_deregister_market

serum3_edit_market

serum3_liq_force_cancel_orders

serum3_place_order

serum3_register_market

Serum

serum3_settle_funds

Deprecated instruction that used to settles all free funds from the OpenOrders account into the MangoAccount.

serum3_settle_funds_v2

Like Serum3SettleFunds, but fees_to_dao determines if referrer rebates are considered fees or are credited to the MangoAccount.

stub_oracle_close

stub_oracle_create

stub_oracle_set

stub_oracle_set_test

token_add_bank

token_conditional_swap_cancel

token_conditional_swap_create

token_conditional_swap_create_v2

token_conditional_swap_trigger

token_deposit

token_deposit_into_existing

token_deregister

token_edit

token_force_close_borrows_with_token

token_liq_bankruptcy

token_liq_with_token

token_register

token_register_trustless

token_update_index_and_rate

token_withdraw