Struct mango_v4::state::DynamicAccount

pub struct DynamicAccount<Header, Fixed, Dynamic> {
    pub header: Header,
    pub fixed: Fixed,
    pub dynamic: Dynamic,
}

Fields
header: Headerfixed: Fixeddynamic: Dynamic
Implementations

impl DynamicAccount<MangoAccountDynamicHeader, MangoAccountFixed, Vec<u8, Global>>

pub fn from_bytes(bytes: &[u8]) -> Result<Self>

impl<'a> DynamicAccount<MangoAccountDynamicHeader, &'a MangoAccountFixed, &'a [u8]>

pub fn from_bytes(bytes: &'a [u8]) -> Result<Self>

impl<Header: DerefOrBorrow<MangoAccountDynamicHeader>, Fixed: DerefOrBorrow<MangoAccountFixed>, Dynamic: DerefOrBorrow<[u8]>> DynamicAccount<Header, Fixed, Dynamic>

pub fn header_version(&self) -> &u8

pub fn token_position_and_raw_index(
&self,
token_index: TokenIndex
) -> Result<(&TokenPosition, usize)>

Returns

the position

the raw index into the token positions list (for use with get_raw/deactivate)

pub fn token_position(&self, token_index: TokenIndex) -> Result<&TokenPosition>

pub fn token_position_by_raw_index(
&self,
raw_index: usize
) -> Result<&TokenPosition>

pub fn all_token_positions(&self) -> impl Iterator<Item = &TokenPosition> + '_

pub fn active_token_positions(
&self
) -> impl Iterator<Item = &TokenPosition> + '_

pub fn serum3_orders(
&self,
market_index: Serum3MarketIndex
) -> Result<&Serum3Orders>

pub fn serum3_orders_by_raw_index(
&self,
raw_index: usize
) -> Result<&Serum3Orders>

pub fn all_serum3_orders(&self) -> impl Iterator<Item = &Serum3Orders> + '_

pub fn active_serum3_orders(&self) -> impl Iterator<Item = &Serum3Orders> + '_

pub fn perp_position(
&self,
market_index: PerpMarketIndex
) -> Result<&PerpPosition>

pub fn perp_position_by_raw_index(
&self,
raw_index: usize
) -> Result<&PerpPosition>

pub fn all_perp_positions(&self) -> impl Iterator<Item = &PerpPosition>

pub fn active_perp_positions(&self) -> impl Iterator<Item = &PerpPosition>

pub fn perp_order_by_raw_index(&self, raw_index: usize) -> Result<&PerpOpenOrder>

pub fn all_perp_orders(&self) -> impl Iterator<Item = &PerpOpenOrder>

pub fn perp_next_order_slot(&self) -> Result<usize>

pub fn perp_find_order_with_client_order_id(
&self,
market_index: PerpMarketIndex,
client_order_id: u64
) -> Option<&PerpOpenOrder>

pub fn perp_find_order_with_order_id(
&self,
market_index: PerpMarketIndex,
order_id: u128
) -> Option<&PerpOpenOrder>

pub fn being_liquidated(&self) -> bool

pub fn token_conditional_swap_by_index(
&self,
index: usize
) -> Result<&TokenConditionalSwap>

pub fn token_conditional_swap_by_id(
&self,
id: u64
) -> Result<(usize, &TokenConditionalSwap)>

pub fn all_token_conditional_swaps(
&self
) -> impl Iterator<Item = &TokenConditionalSwap>

pub fn active_token_conditional_swaps(
&self
) -> impl Iterator<Item = &TokenConditionalSwap>

pub fn token_conditional_swap_free_index(&self) -> Result<usize>

pub fn borrow(&self) -> MangoAccountRef<'_>

impl<Header: DerefOrBorrowMut<MangoAccountDynamicHeader> + DerefOrBorrow<MangoAccountDynamicHeader>, Fixed: DerefOrBorrowMut<MangoAccountFixed> + DerefOrBorrow<MangoAccountFixed>, Dynamic: DerefOrBorrowMut<[u8]> + DerefOrBorrow<[u8]>> DynamicAccount<Header, Fixed, Dynamic>

pub fn borrow_mut(&mut self) -> MangoAccountRefMut<'_>

pub fn token_position_mut(
&mut self,
token_index: TokenIndex
) -> Result<(&mut TokenPosition, usize)>

Returns

the position

the raw index into the token positions list (for use with get_raw/deactivate)

pub fn token_position_mut_by_raw_index(
&mut self,
raw_index: usize
) -> &mut TokenPosition

pub fn ensure_token_position(
&mut self,
token_index: TokenIndex
) -> Result<(&mut TokenPosition, usize, usize)>

Creates or retrieves a TokenPosition for the token_index. Returns:

the position

the raw index into the token positions list (for use with get_raw)

the active index, for use with FixedOrderAccountRetriever

pub fn deactivate_token_position(&mut self, raw_index: usize)

pub fn deactivate_token_position_and_log(
&mut self,
raw_index: usize,
mango_account_pubkey: Pubkey
)

pub fn token_decrement_dust_deactivate(
&mut self,
bank: &mut Bank,
now_ts: u64,
mango_account_pubkey: Pubkey
) -> Result<()>

Decrements the in_use_count for the token position for the bank.

If it goes to 0, the position may be dusted (if between 0 and 1 native tokens) and closed.

pub fn serum3_orders_mut_by_raw_index(
&mut self,
raw_index: usize
) -> &mut Serum3Orders

pub fn create_serum3_orders(
&mut self,
market_index: Serum3MarketIndex
) -> Result<&mut Serum3Orders>

pub fn deactivate_serum3_orders(
&mut self,
market_index: Serum3MarketIndex
) -> Result<()>

pub fn serum3_orders_mut(
&mut self,
market_index: Serum3MarketIndex
) -> Result<&mut Serum3Orders>

pub fn perp_position_mut_by_raw_index(
&mut self,
raw_index: usize
) -> &mut PerpPosition

pub fn perp_order_mut_by_raw_index(
&mut self,
raw_index: usize
) -> &mut PerpOpenOrder

pub fn perp_position_mut(
&mut self,
market_index: PerpMarketIndex
) -> Result<&mut PerpPosition>

pub fn ensure_perp_position(
&mut self,
perp_market_index: PerpMarketIndex,
settle_token_index: TokenIndex
) -> Result<(&mut PerpPosition, usize)>

pub fn deactivate_perp_position(
&mut self,
perp_market_index: PerpMarketIndex,
settle_token_index: TokenIndex
) -> Result<()>

pub fn deactivate_perp_position_and_log(
&mut self,
perp_market_index: PerpMarketIndex,
settle_token_index: TokenIndex,
mango_account_pubkey: Pubkey
) -> Result<()>

pub fn add_perp_order(
&mut self,
perp_market_index: PerpMarketIndex,
side: Side,
order_tree: BookSideOrderTree,
order: &LeafNode,
client_order_id: u64
) -> Result<()>

pub fn remove_perp_order(&mut self, slot: usize, quantity: i64) -> Result<()>

pub fn execute_perp_maker(
&mut self,
perp_market_index: PerpMarketIndex,
perp_market: &mut PerpMarket,
fill: &FillEvent,
group: &Group
) -> Result<()>

pub fn execute_perp_taker(
&mut self,
perp_market_index: PerpMarketIndex,
perp_market: &mut PerpMarket,
fill: &FillEvent
) -> Result<()>

pub fn token_conditional_swap_mut_by_index(
&mut self,
index: usize
) -> Result<&mut TokenConditionalSwap>

pub fn free_token_conditional_swap_mut(
&mut self
) -> Result<&mut TokenConditionalSwap>

pub fn check_health_pre(&mut self, health_cache: &HealthCache) -> Result<I80F48>

pub fn check_health_pre_checks(
&mut self,
health_cache: &HealthCache,
pre_init_health: I80F48
) -> Result<()>

pub fn check_health_post(
&mut self,
health_cache: &HealthCache,
pre_init_health: I80F48
) -> Result<I80F48>

pub fn check_health_post_checks(
&mut self,
pre_init_health: I80F48,
post_init_health: I80F48
) -> Result<()>

pub fn check_liquidatable(
&mut self,
health_cache: &HealthCache
) -> Result<CheckLiquidatable>

pub fn resize_dynamic_content(
&mut self,
new_token_count: u8,
new_serum3_count: u8,
new_perp_count: u8,
new_perp_oo_count: u8,
new_token_conditional_swap_count: u8
) -> Result<()>